- Move duplicate methods into base class - Move FrozenDict to Utils file - Update test to adhere to new error generation order
