+++ tdrs-backend/tdpservice/parsers/tdr_parser.py
@@ -125,8 +125,7 @@ class TanfDataReportParser(BaseParser):
                     self.unsaved_records.add_record(
                         current_case_id, (record, schema.model), self.current_row_num
                     )
-                    if should_remove:
-                        self.serialized_cases.add(case_id_to_remove)
+                    self.add_case_to_remove(should_remove, case_id_to_remove)
 
                     self.dfs.total_number_of_records_in_file += 1
 
@@ -160,16 +159,14 @@ class TanfDataReportParser(BaseParser):
             return
 
         should_remove = self.validate_case_consistency()
-        if should_remove:
-            self.serialized_cases.add(case_id_to_remove)
+        self.add_case_to_remove(should_remove, current_case_id)
 
         # Only checking "all_created" here because records remained cached if bulk create fails.
         # This is the last chance to successfully create the records.
         all_created = self.bulk_create_records(self.header_count, flush=True)
 
-        # Order matters. If a case had case consistency errors and duplicate errors, only the case consistency errors are reflected. We want to keep this order
-        # because that is what happens for the in-memory case now also. I.e. if a case is removed from the in memory cache it won't generate duplicate errors
-        # if they exist.
+        # Order matters. To ensure we also catch duplicate records, we check for them before deleting cases with cat4
+        # Errors.
         self._delete_exact_dups()
         self._delete_partial_dups()
         self._delete_serialized_cases()
@@ -201,126 +198,6 @@ class TanfDataReportParser(BaseParser):
         self.dfs.save()
 
         return
-
-    def _delete_serialized_cases(self):
-        from django.db.models import Q
-
-        if len(self.serialized_cases):
-            cases = Q()
-            for case in self.serialized_cases:
-                cases |= Q(**case)
-            for schemas in self.schema_manager.schema_map.values():
-                schema = schemas[0]
-                num_deleted, _ = schema.model.objects.filter(
-                    cases, datafile=self.datafile
-                ).delete()
-                self.dfs.total_number_of_records_created -= num_deleted
-
-    def _delete_exact_dups(self):
-        """"""
-        from django.db.models import Count, Q
-
-        duplicate_error_generator = self.error_generator_factory.get_generator(
-            ErrorGeneratorType.DYNAMIC_ROW_CASE_CONSISTENCY, None
-        )
-        for schemas in self.schema_manager.schema_map.values():
-            schema = schemas[0]
-
-            fields = [f.name for f in schema.fields if f.name != "BLANK"]
-            dups = (
-                schema.model.objects.filter(datafile=self.datafile)
-                .values(*fields)
-                .annotate(row_count=Count("line_number", distinct=True))
-                .filter(row_count__gt=1)
-            )
-            dup_errors = []
-            print(f"Count: {schema.model.objects.count()}")
-            for d in dups:
-                d.pop("row_count", None)
-                dup_records = schema.model.objects.filter(**d).order_by("line_number")
-                record = dup_records.first()
-                for dup in dup_records[1:]:
-                    record_type = d.get("RecordType", None)
-                    generator_args = ErrorGeneratorArgs(
-                        record=record,
-                        schema=schema,
-                        error_message=f"Duplicate record detected with record type {record_type} at line {dup.line_number}. Record is a duplicate of the record at line number {record.line_number}.",
-                        fields=schema.fields,
-                        row_number=record.line_number,
-                    )
-                    # Perform Error Generation
-                    dup_errors.append(
-                        duplicate_error_generator(generator_args=generator_args)
-                    )
-                num_deleted, _ = dup_records.delete()
-                self.dfs.total_number_of_records_created -= num_deleted
-
-            self.unsaved_parser_errors[None] = (
-                self.unsaved_parser_errors.get(None, []) + dup_errors
-            )
-
-    def __get_partial_dup_error_msg(
-        self, schema, record_type, curr_line_number, existing_line_number
-    ):
-        """Generate partial duplicate error message with friendly names."""
-        field_names = schema.get_partial_hash_members_func()
-        err_msg = (
-            f"Partial duplicate record detected with record type "
-            f"{record_type} at line {curr_line_number}. Record is a partial duplicate of the "
-            f"record at line number {existing_line_number}. Duplicated fields causing error: "
-        )
-        for i, name in enumerate(field_names):
-            field = schema.get_field_by_name(name)
-            item_and_name = f"Item {field.item} ({field.friendly_name})"
-            if i == len(field_names) - 1 and len(field_names) != 1:
-                err_msg += f"and {item_and_name}."
-            elif len(field_names) == 1:
-                err_msg += f"{item_and_name}."
-            else:
-                err_msg += f"{item_and_name}, "
-        return err_msg
-
-    def _delete_partial_dups(self):
-        from django.db.models import Count, Q
-
-        duplicate_error_generator = self.error_generator_factory.get_generator(
-            ErrorGeneratorType.DYNAMIC_ROW_CASE_CONSISTENCY, None
-        )
-        for schemas in self.schema_manager.schema_map.values():
-            schema = schemas[0]
-            fields = schema.get_partial_hash_members_func()
-            dups = (
-                schema.model.objects.filter(datafile=self.datafile)
-                .values(*fields)
-                .annotate(row_count=Count("line_number", distinct=True))
-                .filter(row_count__gt=1)
-            )
-            dup_errors = []
-            for d in dups:
-                d.pop("row_count", None)
-                dup_records = schema.model.objects.filter(**d).order_by("line_number")
-                record = dup_records.first()
-                for dup in dup_records[1:]:
-                    record_type = d.get("RecordType", None)
-                    generator_args = ErrorGeneratorArgs(
-                        record=record,
-                        schema=schema,
-                        error_message=self.__get_partial_dup_error_msg(
-                            schema, record_type, dup.line_number, record.line_number
-                        ),
-                        fields=schema.fields,
-                        row_number=record.line_number,
-                    )
-                    # Perform Error Generation
-                    dup_errors.append(
-                        duplicate_error_generator(generator_args=generator_args)
-                    )
-                num_deleted, _ = dup_records.delete()
-                self.dfs.total_number_of_records_created -= num_deleted
-
-            self.unsaved_parser_errors[None] = (
-                self.unsaved_parser_errors.get(None, []) + dup_errors
-            )
 
     def _validate_header(self):
         """Validate header and header fields."""
@@ -482,8 +359,3 @@ class TanfDataReportParser(BaseParser):
         if settings.GENERATE_TRAILER_ERRORS:
             self.unsaved_parser_errors.update({"trailer": trailer_errors})
             self.num_errors += len(trailer_errors)
-
-    def delete_serialized_records(self):
-        """Delete all records that have already been serialized to the DB that have cat4 errors."""
-        duplicate_manager = self.case_consistency_validator.duplicate_manager
-        self._delete_serialized_records(duplicate_manager)
