+++ tdrs-backend/tdpservice/search_indexes/models/mixins.py
@@ -1,7 +1,6 @@
 """Mixin classes."""
 
 from django.db import models
-from django.utils.translation import gettext_lazy as _
 
 
 class LineNumberMixin(models.Model):
